Module orml_traits::currency

Enums

BalanceStatus

Status of funds.

Traits

BasicCurrency

Abstraction over a fungible (single) currency system.

BasicCurrencyExtended

Extended BasicCurrency with additional helper types and methods.

BasicLockableCurrency

A fungible single currency system whose accounts can have liquidity restrictions.

BasicReservableCurrency

A fungible single currency system where funds can be reserved from the user.

MultiCurrency

Abstraction over a fungible multi-currency system.

MultiCurrencyExtended

Extended MultiCurrency with additional helper types and methods.

MultiLockableCurrency

A fungible multi-currency system whose accounts can have liquidity restrictions.

MultiReservableCurrency

A fungible multi-currency system where funds can be reserved from the user.

OnDust

Handler for account which has dust, need to burn or recycle it

TransferAll

Type Definitions

LockIdentifier

An identifier for a lock. Used for disambiguating different locks so that they can be individually replaced or removed.

Attribute Macros

transactional

Execute the annotated function in a new storage transaction.